requirements.txt
fastapi
uvicorn[standard]
jinja2
openenv
ollama
python-dotenv